+++ index.js
@@ -1,7 +1,7 @@
 "use strict";
 
 /*
-This script converts CCDA data in JSON format (originally generated from a Continuity of Care Document (CCD) in 
+This script converts CCDA data in JSON format (originally generated from a Continuity of Care Document (CCD) in
 standard XML/CCDA format) back to XML/CCDA format.
 */
 
@@ -34,6 +34,7 @@ var createContext = (function () {
         } else {
             result.rootId = null;
         }
+        result.preventNullFlavor = options.preventNullFlavor;
 
         return result;
     };